+++ packages/terra-core-docs/src/terra-dev-site/test/paginator/ControlledPaginatorNoPages.test.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import Button from 'terra-button';
 import classNames from 'classnames/bind';
-import ControlledPaginator from '../../../ControlledPaginator';
+import ControlledPaginator from 'terra-paginator/lib/ControlledPaginator';
 import styles from './ControlledPaginatorTestCommon.module.scss';
 
 const cx = classNames.bind(styles);